% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{subsetPrecip}
\alias{subsetPrecip}
\alias{subsetPrecip}
\title{subsetPrecip}
\usage{
subsetPrecip(data)
}
\arguments{
\item{data}{data frame; The weather data frame.}
}
\description{
subsetPrecip
}
\author{
Jason T. Serviss
}
