% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFigures.R
\name{plotFigure}
\alias{plotFigure}
\alias{plotFigure}
\title{plotFigure}
\usage{
plotFigure(figure, ...)
}
\arguments{
\item{figure}{Character; The figure to plot.}

\item{...}{additional arguments to pass on.}
}
\description{
Recreates any figure from scratch.
}
\details{
Include description.
}
\examples{

\dontrun{plotFigure("figure1a")}

}
\author{
Jason T. Serviss
}
